Separate "Host assets" from Angular assets
