+++ src/webpack/plugins/open-host-url/open-host-url.ts
@@ -5,8 +5,8 @@ import {
 } from 'webpack';
 
 import {
-  getAssets
-} from '../../stats-utils';
+  getHostAssets
+} from '../../host-utils';
 
 import {
   SkyuxHostConfig
@@ -37,7 +37,7 @@ export class SkyuxOpenHostURLPlugin {
         return;
       }
 
-      const assets = getAssets(webpackStats.toJson());
+      const assets = getHostAssets(webpackStats.toJson());
       const config: SkyuxHostConfig = {
         localUrl: this.config.localUrl,
         scripts: assets
